feat: Add Orders & Payments Filament resources
Introduce Filament resources for Orders and Payments (resources, pages, tables, infolists and forms) and wire up UI changes to use User::fullName.

Update shopping cart / checkout and payments flow: return ShoppingCartItemResource from add-to-cart, add logging, remove cart clear endpoint and policy, add repository row locking and refine item create/update behavior, and provide a clearPurchasedItems path.

CheckoutService now carries shopping_cart_id, validates product items, builds Stripe line items with proper currency/rounding and errors on missing items. InventoryService and PaymentService adapted to use Collections; PaymentService now injects ShoppingCartService and clears purchased items after payment. Models and migrations updated (Order.shopping_cart_id, ShoppingCartItem fillable fix, migration tweaks). Deleted unused HandleStripeSessionCompleted job.

+++ backend/database/migrations/2026_04_19_040342_create_shopping_cart_items_table.php
@@ -16,10 +16,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('shopping_cart_items', function (Blueprint $table) {
+
             $table->id();
+
             $table->foreignIdFor(ShoppingCart::class)->constrained()->cascadeOnDelete();
+
             $table->foreignIdFor(ProductItem::class)->constrained()->cascadeOnDelete();
-            $table->unsignedBigInteger('quantity');
+
+            $table->unsignedInteger('quantity');
+
             $table->timestamps();
         });
     }
